--- src/main/java/com/banghyang/oauth/domain/OauthMemberRoleType.java
@@ -1,6 +1,0 @@
-package com.banghyang.oauth.domain;
-
-public enum OauthMemberRoleType {
-    USER,
-    ADMIN
-}
