--- src/main/java/com/banghyang/oauth/infra/oauth/kakao/dto/KakaoToken.java
@@ -1,17 +1,0 @@
-package com.banghyang.oauth.infra.oauth.kakao.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoToken(
-        // https://developers.kakao.com/docs/latest/ko/kakaologin/rest-api#request-token
-        String tokenType,
-        String accessToken,
-        String idToken,
-        Integer expiresIn,
-        String refreshToken,
-        Integer refreshTokenExpiresIn,
-        String scope
-) {
-}
