+++ src/main/java/com/banghyang/oauth/kakao/config/KakaoOauthConfig.java
@@ -1,0 +1,13 @@
+package com.banghyang.oauth.kakao.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+// 상수값들을 저장해놓은 properties 파일에서 값들을 가져오는 클래스
+@ConfigurationProperties(prefix = "oauth.kakao")
+public record KakaoOauthConfig(
+        String redirectUri,
+        String clientId,
+        String clientSecret,
+        String[] scope
+) {
+}
