+++ src/main/java/com/banghyang/oauth/kakao/model/dto/KakaoToken.java
@@ -1,0 +1,16 @@
+package com.banghyang.oauth.kakao.model.dto;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class) // SnakeCase 로 Json Data 이름 매핑시킴
+public record KakaoToken(
+        String tokenType,
+        String accessToken,
+        String idToken,
+        Integer expiresIn,
+        String refreshToken,
+        Integer refreshTokenExpiresIn,
+        String scope
+) {
+}
