+++ src/main/java/com/banghyang/oauth/kakao/service/KakaoService.java
@@ -1,0 +1,57 @@
+package com.banghyang.oauth.kakao.service;
+
+import com.banghyang.oauth.kakao.client.KakaoAPIClient;
+import com.banghyang.oauth.kakao.config.KakaoOauthConfig;
+import com.banghyang.oauth.kakao.member.model.dto.KakaoMemberResponse;
+import com.banghyang.oauth.kakao.model.dto.KakaoToken;
+import com.banghyang.oauth.member.dto.OauthMemberResponse;
+import com.banghyang.oauth.member.entity.OauthMember;
+import com.banghyang.oauth.member.repository.OauthMemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.web.util.UriComponentsBuilder;
+
+@Service
+@RequiredArgsConstructor
+public class KakaoService {
+
+    private final KakaoOauthConfig kakaoOauthConfig;
+    private final KakaoAPIClient kakaoAPIClient;
+    private final OauthMemberRepository oauthMemberRepository;
+
+    public String getKakaoAuthCodeRequestUrl() {
+        return UriComponentsBuilder
+                .fromUriString("https://kauth.kakao.com/oauth/authorize")
+                .queryParam("response_type", "code")
+                .queryParam("client_id", kakaoOauthConfig.clientId())
+                .queryParam("redirect_uri", kakaoOauthConfig.redirectUri())
+                .queryParam("scope", String.join(",", kakaoOauthConfig.scope()))
+                .toUriString();
+    }
+
+    public OauthMemberResponse createOauthMemberFromKakao(String authCode) {
+        System.out.println("AuthCode 값 확인 : " + authCode);
+        // AccessToken 발급
+        KakaoToken token = kakaoAPIClient.getToken(
+                "authorization_code",
+                kakaoOauthConfig.clientId(),
+                kakaoOauthConfig.redirectUri(),
+                authCode,
+                kakaoOauthConfig.clientSecret()
+        );
+
+        // 사용자 정보 가져오기
+        KakaoMemberResponse kakaoMemberResponse = kakaoAPIClient
+                .getKakaoMember("bearer " + token.accessToken());
+        System.out.println("AccessToken값 확인 : " + token.accessToken());
+
+        // 사용자 정보를 엔티티로 변환
+        OauthMember oauthMember = kakaoMemberResponse.toEntity();
+
+        // 기존 사용자인지 확인하고 신규 사용자라면 저장
+        OauthMember saved = oauthMemberRepository.findByOauthId(oauthMember.getOauthId())
+                .orElseGet(() -> oauthMemberRepository.save(oauthMember));
+
+        return OauthMemberResponse.from(saved);
+    }
+}
