+++ src/main/java/com/banghyang/oauth/member/dto/OauthId.java
@@ -1,0 +1,31 @@
+package com.banghyang.oauth.member.dto;
+
+import com.banghyang.oauth.member.type.OauthServerType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class OauthId {
+
+    @Column(nullable = false)
+    private String oauthServerId;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false, name = "oauth_server")
+    private OauthServerType oauthServerType;
+
+    public String OauthId() {
+        return oauthServerId;
+    }
+
+    public OauthServerType OauthServer() {
+        return oauthServerType;
+    }
+}
