+++ src/main/java/com/banghyang/oauth/member/type/OauthMemberRoleType.java
@@ -1,0 +1,6 @@
+package com.banghyang.oauth.member.type;
+
+public enum OauthMemberRoleType {
+    USER,
+    ADMIN
+}
